% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtriangular.R
\name{dtriangular}
\alias{dtriangular}
\title{Triangular Distribution}
\usage{
dtriangular(x, min, peak, max, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{min}{lower bound of distribution}

\item{peak}{peak of triangle density}

\item{max}{upper bound of distribution}

\item{log}{logical; if TRUE, probabilities p are given as log(p)}
}
\description{
Distribution with a triangular density function.
}
\examples{
curve(dtriangular(x, min = .2, peak = .6, max = 1.3),
      from = 0, to = 2, n = 301)

plot(prior("triangular", c(.2, .6, 1.3)), 0, 2)
}
