% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ordered.R
\name{meta_ordered}
\alias{meta_ordered}
\title{Meta-Analysis with Order-Constrained Study Effects}
\usage{
meta_ordered(y, SE, labels, data, d = prior("norm", c(mean = 0, sd =
  0.3), lower = 0), tau = prior("invgamma", c(shape = 1, scale = 0.15)),
  prior = c(1, 1, 1, 1), logml = "integrate", summarize = "stan",
  ci = 0.95, rel.tol = .Machine$double.eps^0.3, ...)
}
\arguments{
\item{y}{effect size per study. Can be provided as (1) a numeric vector, (2)
the quoted or unquoted name of the variable in \code{data}, or (3) a
\code{\link[stats]{formula}} to include discrete or continuous moderator
variables.}

\item{SE}{standard error of effect size for each study. Can be a numeric
vector or the quoted or unquoted name of the variable in \code{data}}

\item{labels}{optional: character values with study labels. Can be a
character vector or the quoted or unquoted name of the variable in
\code{data}}

\item{data}{data frame containing the variables for effect size \code{y},
standard error \code{SE}, \code{labels}, and moderators per study.}

\item{d}{\code{prior} distribution on the average effect size \eqn{d}. The
prior probability density function is defined via \code{\link{prior}}.}

\item{tau}{\code{prior} distribution on the standard deviation \eqn{\tau} of
the study effect sizes in a random-effects meta-analysis (i.e., the SD of
\eqn{d} across studies). A (nonnegative) prior probability density function
is defined via \code{\link{prior}}.}

\item{prior}{prior probabilities over models (possibly unnormalized) in the
order \code{c(fixed_H0, fixed_H1, ordered_H1, random_H1)}. Note that the
model \code{random_H0} is not included in the comparison.}

\item{logml}{how to estimate the log-marginal likelihood: either by numerical
integration (\code{"integrate"}) or by bridge sampling using MCMC/Stan
samples (\code{"stan"}). To obtain high precision with \code{logml="stan"},
many MCMC samples are required (e.g., \code{iter=10000, warmup=1000}).}

\item{summarize}{how to estimate parameter summaries (mean, median, SD,
etc.): Either by numerical integration (\code{summarize = "integrate"}) or
based on MCMC/Stan samples (\code{summarize = "stan"}).}

\item{ci}{probability for the credibility/highest-density intervals.}

\item{rel.tol}{relative tolerance used for numerical integration using
\code{\link[stats]{integrate}}. Use \code{rel.tol=.Machine$double.eps} for
maximal precision (however, this might be slow).}

\item{...}{further arguments passed to \code{rstan::sampling} (see
\code{\link[rstan]{stanmodel-method-sampling}}). For instance:
\code{warmup=500}, \code{chains=4}, \code{control=list(adapt_delta=.95)}).}
}
\description{
Computes the Bayes factor for the hypothesis that the true study effects in a
random-effects meta-anaysis are all positive or negative.
}
\details{
The Bayes factor for the order-constrained model is computed using the
encompassing Bayes factor.
}
\examples{
data(towels)
debug(meta_ordered)
mo <- meta_ordered(logOR, SE, study, towels,
                   d = prior("norm", c(mean=0, sd=.3), lower=0),
                   tau = prior("invgamma", c(shape = 1, scale = 0.15)))
mo
plot_posterior(mo, "d")
}
\references{
Haaf, J. M., & Rouder, J. N. (in press). Some do and some don’t? Accounting for variability of individual difference structures. Psychonomic Bulletin & Review. doi: 10.3758/s13423-018-1522-x
}
\seealso{
\link{meta_random}
}
