% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_stan.R
\name{meta_stan}
\alias{meta_stan}
\title{Bayesian Meta-Analysis Using Stan}
\usage{
meta_stan(y, SE, labels, data, model = c("random", "fixed"), d = c(1, 0,
  0.3, -Inf, Inf), tau = c(1, 0, 0.5, 0, Inf), scale = 0.5, show = FALSE,
  ...)
}
\arguments{
\item{y}{effect size per study. Can be provided as (1) a numeric vector,
(2) the quoted or unquoted name of the variable in \code{data}, or
(3) a \code{\link[stats]{formula}} to include discrete or continuous moderator
variables (only for \code{meta_stan}).}

\item{SE}{standard error of effect size for each study. Can be a numeric vector
or the quoted or unquoted name of the variable in \code{data}}

\item{labels}{optional: character values with study labels. Can be a character vector
or the quoted or unquoted name of the variable in \code{data}}

\item{data}{data frame containing the variables for effect size \code{y},
standard error \code{SE}, \code{labels}, and moderators per study.}

\item{model}{whether to fit a \code{"random"} or \code{"fixed"} effects meta-analysis
(can be abbreviated by \code{"r"} and \code{"f"}).}

\item{d}{parameters defining the prior on the mean effect size d
(parameters of a truncated t distribution: df, location, scale, lower, upper).}

\item{tau}{parameters defining the prior on the heterogeneity tau
(parameters of a truncated t distribution: df, location, scale, lower, upper).}

\item{scale}{scale parameter(s) of the JZS prior for the moderators.
If a scalar is provided, the same prior is used for all JZS blocks.
If a vector is provided, the scale values are assined in the same order as
defined in the formula \code{y}.}

\item{show}{whether to print Stan output.}

\item{...}{further arguments passed to \code{\link[rstan]{sampling}}.}
}
\description{
Uses Stan to draw posterior samples of a fixed-effects or random-effects meta-analysis
with or without moderator variables.
}
