% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Plot Posterior Distribution}
\usage{
plot_posterior(meta, parameter = "d", from = 0, to = 1,
  summary = c("Mean", "HPD"), ...)
}
\arguments{
\item{meta}{fitted meta-analysis model}

\item{parameter}{only for random-effects model: whether to plot \code{"d"} or \code{"tau"}}

\item{from}{lower limit of the x-axis}

\item{to}{upper limit of the x-axis}

\item{summary}{character vector with two values: first, either \code{"Mean"} or \code{"Median"}; and second, either highest-probability-density interval (\code{"HPD"}) or the quantile interval (\code{"quantile"}; function resets margin to default after plotting)}

\item{...}{arguments passed to \code{\link[graphics]{plot}}}
}
\description{
Plot Posterior Distribution
}
\seealso{
\link{meta_default}, \link{meta_bma}, \link{meta_fixed}, \link{meta_random}
}
