% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.R
\name{predictive}
\alias{predictive}
\title{Predicted Bayes Factor for a New Study}
\usage{
predictive(meta, SE, resample = 100)
}
\arguments{
\item{meta}{model-averaged meta-analysis (fitted with \code{\link{meta_bma}} or \code{\link{meta_default}})}

\item{SE}{expected standard error of future study. For instance, SE = 1/sqrt(N) for standardized effect sizes and N = sample size)}

\item{resample}{number of simulated Bayes factors}
}
\description{
How much can be learned by an additional study? To judge this, this function samples the distribution of predicted Bayes factors for a new study given the current evidence.
}
