% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior}
\alias{prior}
\title{Prior Distributions}
\usage{
prior(prior, log = FALSE)
}
\arguments{
\item{prior}{a named list with the two arguments \code{name} (character) and \code{par} (numeric vector)}

\item{log}{whether to return log density}
}
\value{
a log-density function with a single argument for the values to be evaluated (used for integration internally)
}
\description{
Supported prior distributions for mean effect size \eqn{d} and heterogeneity of effect sizes \eqn{\tau}
}
\details{
\itemize{
\item \code{"norm"}: Normal distribution with \code{par = c(mean, sd)} (see \code{\link[stats]{dnorm}})
\item \code{"halfnorm"}: Half-normal distribution (positive) with \code{par = c(mean, sd)} (see \code{\link[LaplacesDemon]{dhalfnorm}})
\item \code{"truncnorm"}: Truncated normal distribution with \code{par = c(min, max, mean, sd)}  (see \code{\link{dtruncnorm}})
\item \code{"scaledt"}: Scaled Student-t with \code{par = c(mean, sigma, nu)} (see \code{\link[LaplacesDemon]{dst}})
\item \code{"halft"}: Half-normal Student-t with \code{par = c(scale, df)} (see \code{\link[LaplacesDemon]{dhalft}})
\item \code{"cauchy"}: Cauchy distribution with \code{par = scale} (see \code{\link[stats]{dcauchy}})
\item \code{"halfcauchy"}: Half-Cauchy distribution (positive) with \code{par = scale}  (see \code{\link[LaplacesDemon]{dhalfcauchy}})
\item \code{"beta"}: Beta distribution with \code{par = c(alpha, beta)} (see \code{\link[stats]{dbeta}})
}
}
\examples{
### Half-Normal Distribution
p1 <- prior(list(name = "halfnorm",
                 par = c(0, .3)))
p1(c(-1,1,3))
plot(p1)

### Half-Cauchy Distribution
p2 <- prior(list(name = "halfcauchy",
                 par = .3))
plot(p2, 0, 3)
}
