% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{bma}
\alias{bma}
\title{Compute Inclusion Bayes Factor}
\usage{
bma(logml, include = 1, prior = 1)
}
\arguments{
\item{logml}{a vector with log marginal likelihoods. Alternatively, a list with fitted meta-analysis models (see \code{\link{meta_random}} and \code{\link{meta_fixed}})}

\item{include}{integer vector which models to include in inclusion posterior. If only two marginal likelihoods/meta-analyses are supplied, the Bayes factor BF_{M1,M2} is computed by default.}

\item{prior}{prior probabilities over models (possibly unnormalized). For instance, if the first model is as likely as models 2, 3 and 4 together: \code{prior = c(3,1,1,1)}}
}
\description{
Compute Inclusion Bayes Factor
}
\examples{
x <- rnorm(50)
# M1: x ~ Normal(0,1)
logm1 <- sum(dnorm(x, log = TRUE))
# M2: x ~ Normal(1,1)
logm2 <-sum(dnorm(x, 2, log = TRUE))
# M3: x ~ Student-t(df=2)
logm3 <-sum(dt(x, df=1, log = TRUE))

# BF: Correct (M1) vs. misspecified (M2, M3)
bma(c(logm1, logm2, logm3), include = 1)
}
