% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{bma}
\alias{bma}
\title{Bayesian Model Averaging}
\usage{
bma(meta, prior = 1, parameter = "d", summarize = "integrate",
  ci = 0.95, rel.tol = .Machine$double.eps^0.5)
}
\arguments{
\item{meta}{list of meta-analysis models (fitted via \code{\link{meta_random}} or \code{\link{meta_fixed}})}

\item{prior}{prior probabilities over models (possibly unnormalized). For instance, if the first model is as likely as models 2, 3 and 4 together: \code{prior = c(3,1,1,1)}. The default is a discrete uniform distribution over models.}

\item{parameter}{eiher the mean effect \code{"d"} or the heterogeneity across studies \code{"tau"}}

\item{summarize}{how to estimate parameter summaries (mean, median, SD, etc.):
Either by numerical integration (\code{summarize = "integrate"}) or based
on MCMC/Stan samples (\code{summarize = "stan"}).}

\item{ci}{probability for the credibility/highest-density intervals.}

\item{rel.tol}{relative tolerance used for numerical integration using
\code{\link[stats]{integrate}}. Use \code{rel.tol=.Machine$double.eps}
for maximal precision (however, this might be slow).}
}
\description{
Model averaging of meta-analysis models according to their posterior model probability.
}
\examples{
data(towels)
fix1 <- meta_fixed(logOR, SE, study, towels,
                   d = prior("norm", c(mean=0, sd=.2), lower=0))
fix2 <- meta_fixed(logOR, SE, study, towels,
                   d = prior("beta", c(shape1=1, shape2=1), upper = 2))
fix3 <- meta_fixed(logOR, SE, study, towels,
                   d = prior("custom", function(x) x^2, 0, 1))

averaged <- bma(list(Halfnormal = fix1, Uniform = fix2,
                     Quadratic = fix3))
averaged
plot_posterior(averaged)
plot_forest(averaged, mar = c(4.5,20,4,.3))
}
