% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtruncnorm.R
\name{dtruncnorm}
\alias{dtruncnorm}
\title{Truncated Normal Distribution}
\usage{
dtruncnorm(x, a = -Inf, b = Inf, mean = 0, sd = 1, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{a}{lower bound}

\item{b}{upper bound}

\item{mean}{vector of means}

\item{sd}{vector of standard deviations}

\item{log}{if TRUE, probabilities are given as log(p)}
}
\description{
Normal distribution truncated to the interval [a,b]
}
\examples{
curve(dtruncnorm(x,-.5,3.5, 1,1), -2,4, n = 501)
}
