% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_default.R
\name{meta_default}
\alias{meta_default}
\title{Defaults for Model Averaging in Meta-Analysis}
\usage{
meta_default(y, V, field = "psychology", effect = "ttest", ...)
}
\arguments{
\item{y}{mean in each study}

\item{V}{variance in each study}

\item{field}{either\code{"psychology"} or \code{"medicine"}}

\item{effect}{the type of effect size: either means (\code{"ttest"}), log-odds ratios (\code{"logOR"}) or (Fisher's z-transformed) correlations (\code{"corr"})}

\item{...}{further arguments passed to \code{\link{meta_bma}}}
}
\description{
Wrapper with default prior for Bayesian meta-analysis based on a literature review.
}
\details{
Default prior distributions can be plotted using \code{\link{plot_default}}.

For \code{field = "psychology"}, the following defaults are used:
\itemize{
\item \code{effect = "ttest"}: Half-normal with SD=0.3 on mean effect and half-Cauchy with scale=.5 on variance of effects.
\item \code{effect = "logOR"}: Half-normal with SD=0.3 on mean effect and half-Cauchy with scale=.5 on variance of effects.
\item \code{effect = "corr"}: Half-normal with SD=0.3 on mean effect and half-Cauchy with scale=.5 on variance of effects.
}

For \code{field = "medicine"}, the following defaults are used:
\itemize{
\item \code{effect = "ttest"}: Half-normal with SD=0.3 on mean effect and half-Cauchy with scale=.5 on variance of effects.
\item \code{effect = "logOR"}: Half-normal with SD=0.3 on mean effect and half-Cauchy with scale=.5 on variance of effects.
\item \code{effect = "corr"}: Half-normal with SD=0.3 on mean effect and half-Cauchy with scale=.5 on variance of effects.
}
}
\seealso{
\link{meta_bma}
}
