% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_fixed.R
\name{meta_fixed}
\alias{meta_fixed}
\title{Bayesian Fixed-Effects Meta-Analysis}
\usage{
meta_fixed(y, V, d = "halfnorm", d.par = c(mean = 0, sd = 0.3),
  marginal = "integrate", ...)
}
\arguments{
\item{y}{mean in each study}

\item{V}{variance in each study}

\item{d}{type of prior for mean effect \eqn{d} (see \code{\link{log_prior}})}

\item{d.par}{prior parameters for \eqn{d}}

\item{marginal}{how to integrate marginal likelihood (\code{"bridge"} or \code{"integrate"})}

\item{...}{arguments passed to \link[runjags]{run.jags} (e.g., MCMC parameters such as \code{sample}, \code{burnin}, \code{n.chains}, \code{thin} or \code{method="parallel"})}
}
\description{
Bayesian Fixed-Effects Meta-Analysis
}
\examples{
data(towels)
### Bayesian Fixed-Effects Meta-Analysis (H1: d>0 Cauchy)
mf <- meta_fixed(towels$logOR, towels$variance,
                 d = "halfnorm", d.par = c(0, .3),
                 sample=300, n.chains = 2,
                 marginal = "integrate")
mf
}

