% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_random.R
\name{meta_random}
\alias{meta_random}
\title{Bayesian Random-Effects Meta-Analysis}
\usage{
meta_random(y, SE, labels, d = "norm", d.par = c(0, 0.3),
  tau = "halfcauchy", tau.par = 0.5, sample = 10000,
  summarize = "jags", rel.tol = .Machine$double.eps^0.5, ...)
}
\arguments{
\item{y}{mean in each study}

\item{SE}{standard error in each study}

\item{labels}{optional: character values with study labels}

\item{d}{a character value specifying the prior family for the mean effect
size \eqn{d} (see \code{\link{prior}}). Alternatively, \code{d} can be a
\code{prior}-function as returned by \code{\link{prior}} (in which case \code{d.par}
is ignored).}

\item{d.par}{prior parameters for \eqn{d}.}

\item{tau}{a character specifying the prior family for the standard deviation \eqn{\tau}
of the study effects in a random-effects meta-analysis (i.e., the SD of d
across studies; see \code{\link{prior}}). Alternatively, \code{tau} can be a
\code{prior}-function as returned by \code{\link{prior}} (in which case \code{tau.par}
is ignored).}

\item{tau.par}{prior parameters for \eqn{\tau}.}

\item{sample}{number of samples in JAGS after burn-in and thinning (see
\code{\link[runjags]{run.jags}}). Samples are used to get posterior
estimates for each study effect (which will show shrinkage).
Only works for priors defined in \code{\link{prior}}.}

\item{summarize}{whether and to compute parameter summaries (mean, median, SD,
95\% quantile interval, HPD interval). If \code{summarize = "integrate"},
numerical integration is used  (which is precise but can require some seconds
of computing time and be unstable), \code{summarize = "jags"} summarizes the JAGS samples,
and \code{summarize = "none"} suppresses parameter summaries.}

\item{rel.tol}{relative tolerance used for numerical integration using \code{\link[stats]{integrate}}.
Use \code{rel.tol=.Machine$double.eps} for maximal precision.}

\item{...}{arguments passed to \link[runjags]{run.jags} (e.g., MCMC parameters
such as \code{sample}, \code{burnin}, \code{n.chains}, \code{thin} or \code{method="parallel"})}
}
\description{
Runs a Bayesian meta-analysis assuming that the effect size \eqn{d} varies across studies with standard deviation \eqn{\tau} (i.e., a random-effects analysis).
}
\examples{
data(towels)
### Bayesian Random-Effects Meta-Analysis
mr <- meta_random(towels$logOR, towels$SE, towels$study,
                  d = "norm", d.par = c(0,.3),
                  tau = "halfcauchy", tau.par = .5,
                  rel.tol = .Machine$double.eps^.15,
                  sample = 0, summarize = "none")
                  # (no summary: only for CRAN checks)
mr
plot_posterior(mr)
}
