% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_random.R
\name{meta_random}
\alias{meta_random}
\title{Bayesian Random-Effects Meta-Analysis}
\usage{
meta_random(y, V, d = "norm", d.par = c(0, 0.3), tau = "halfcauchy",
  tau.par = 0.5, marginal = "integrate", ...)
}
\arguments{
\item{y}{mean in each study}

\item{V}{variance in each study}

\item{d}{type of prior for mean effect \eqn{d} (see \code{\link{log_prior}})}

\item{d.par}{prior parameters for \eqn{d}}

\item{tau}{type of prior for study-variance \eqn{\tau} in random-effects meta-analysis (i.e., the variance of d across studies; see \code{\link{log_prior}})}

\item{tau.par}{prior parameters for \eqn{\tau}}

\item{marginal}{how to integrate marginal likelihood (\code{"bridge"} or \code{"integrate"})}

\item{...}{arguments passed to \link[runjags]{run.jags} (e.g., MCMC parameters such as \code{sample}, \code{burnin}, \code{n.chains}, \code{thin} or \code{method="parallel"})}
}
\description{
Bayesian Random-Effects Meta-Analysis
}
\examples{
data(towels)
### Bayesian Random-Effects Meta-Analysis (H0: No effect, d=0)
mr <- meta_random(towels$logOR, towels$variance,
                  d = "0",  ### random-efects H0 ###
                  tau = "halfcauchy",tau.par = .5,
                  sample=300, n.chains = 2,
                  marginal = "integrate")
mr
}

