% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_metaBMA-package.R
\docType{package}
\name{metaBMA-package}
\alias{metaBMA}
\alias{metaBMA-package}
\title{metaBMA: Bayesian Model Averaging for Random and Fixed Effects Meta-Analysis}
\description{
metaBMA: Bayesian Model Averaging for Random and Fixed Effects Meta-Analysis
}
\references{
Heck, D. W. & Gronau, Q. F. (2017). metaBMA: Bayesian Model Averaging for Random and Fixed Effects Meta-Analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/danheck/metaBMA}
}

}
\author{
Daniel W. Heck & Quentin F. Gronau
}
