% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_default.R
\name{plot_default}
\alias{plot_default}
\title{Plot Default Priors}
\usage{
plot_default(field = "psychology", effect = "d", ...)
}
\arguments{
\item{field}{either\code{"psychology"} or \code{"medicine"}
(uses partial matching, so \code{"p"} and "\code{"m"} are sufficient)}

\item{effect}{the type of effect size: either
Cohen's d (\code{"d"}),
Pearson correlations (\code{"r"}),
Fisher's z-transformed correlations (\code{"z"}),
or log-odds ratios (\code{"logOR"}).}

\item{...}{further arguments passed to \code{\link[graphics]{plot}} (e.g., \code{from}, \code{to})}
}
\description{
Plots default priors for the mean effect \eqn{d} and the standard deviation of effects \eqn{tau}.
}
\examples{
plot_default("psychology", "ttest", 0, 2)
plot_default("medicine", "logOR", 0, 2)
}
\seealso{
\code{\link{meta_default}} for details on standard priors.
}
