% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forest.R
\name{plot_forest}
\alias{plot_forest}
\title{Forest Plot for Meta-Analysis}
\usage{
plot_forest(meta, from = 0, to = 1, ...)
}
\arguments{
\item{meta}{fitted meta-analysis model}

\item{from}{extending the lower limit of the x-axis}

\item{to}{extending the upper limit of the x-axis}

\item{...}{arguments passed to \code{\link[graphics]{plot}} (e.g., \code{from}, \code{to})}
}
\description{
Plots estimated effect sizes for all studies.
}
\seealso{
\link{meta_default}, \link{meta_bma}, \link{meta_fixed}, \link{meta_random}
}
