% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Plot Posterior Distribution}
\usage{
plot_posterior(meta, from, to, parameter = "d", ...)
}
\arguments{
\item{meta}{fitted meta-analysis model}

\item{...}{arguments passed to \code{\link[base]{plot}}}
}
\description{
Plot Posterior Distribution
}

