% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_towels.R
\docType{data}
\name{towels}
\alias{towels}
\title{Data: Reuse of Towels in Hotels}
\format{A data frame 2 variables:
\describe{
  \item{\code{logOR}}{log odds ratio of towel reuse}
  \item{\code{variance}}{variance for each study}
}}
\usage{
towels
}
\description{
Dataset xxx
}
\details{
xxx
}
\examples{
data(towels)
head(towels)
}
\references{
Scheibehenne, B., Jamil, T., & Wagenmakers, E.-J. (2016). Bayesian Evidence Synthesis Can Reconcile Seemingly Inconsistent Results: The Case of Hotel Towel Reuse. Psychological Science, 27(7), 1043–1046. https://doi.org/10.1177/0956797616644081
}
\keyword{datasets}
